:recycle: Refactor impl api files
