+++ order/server/src/main/kotlin/io/reyurnible/order/routes/api/ApiRouting.kt
@@ -1,6 +1,5 @@
 package io.reyurnible.order.routes.api
 
-import io.ktor.server.application.*
 import io.ktor.server.routing.*
 import io.reyurnible.order.routes.api.users.usersApi
